Extract rendering helpers to keep files under 1000 lines
Move render_history, format_message, and render_input from
chat_view/mod.rs to chat_view/render_helpers.rs. Move view rendering
logic from line_input/mod.rs to line_input/view_helpers.rs.
